fix(health): move legDetail out of the server-only module — CI's build caught what tsc cannot
The banner is a CLIENT component and lib/ingest/pipeline-health is
`import "server-only"`, so importing the composition from there broke
`next build` while `npx tsc --noEmit` and the entire unit tier stayed green —
the test config even aliases `server-only` to an empty stub. Anything both sides
need now lives in lib/ingest/leg-detail, where neither boundary applies.

Verified the way it should have been the first time: `npm run build` locally,
not just typecheck.
